refactor: separated field components/types, updated imports
