+++ components/builder/editor-pane/field-settings-dialog.tsx
@@ -12,7 +12,7 @@ import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Textarea } from "@/components/ui/textarea";
-import { Field, FieldSettings } from "../field-utils";
+import { Field, FieldSettings } from "@/types/field";
 
 interface FieldSettingsDialogProps {
 	field: Field;
